+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2022-05-20
+
+### Added
+
+- Log warning when core initialize with no cloud project id
+- Add a message in "Link your unity project" popup to inform the user has to sign-up
+
 ## [1.4.0] - 2022-04-29
 
 ### Added